LayoutTests/fast/css-generated-content/pseudo-transition-expected.txt

Transitions on :before and :after pseudo elements should run

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS div.offsetWidth is 52
PASS div.offsetWidth is 20
PASS div.offsetWidth is 12
PASS div.offsetWidth is 52
PASS div.offsetWidth is 20
PASS div.offsetWidth is 12
PASS successfullyParsed is true

TEST COMPLETE
 
